Fix problems with OpenTV processing.

+++ src/epg.h
@@ -343,7 +343,7 @@ struct epg_broadcast
 {
   epg_object_t;                                ///< Parent object
   
-  int                        dvb_eid;          ///< DVB Event ID
+  uint16_t                   dvb_eid;          ///< DVB Event ID
   time_t                     start;            ///< Start time
   time_t                     stop;             ///< End time
 
@@ -373,7 +373,7 @@ struct epg_broadcast
 epg_broadcast_t *epg_broadcast_find_by_time 
   ( struct channel *ch, time_t start, time_t stop, 
     uint16_t eid, int create, int *save );
-epg_broadcast_t *epg_broadcast_find_by_eid ( struct channel *ch, int eid );
+epg_broadcast_t *epg_broadcast_find_by_eid ( struct channel *ch, uint16_t eid );
 epg_broadcast_t *epg_broadcast_find_by_id  ( uint64_t id, struct channel *ch );
 
 /* Mutators */
